+++ src/Exception/InvalidIdentifierException.php
@@ -19,7 +19,7 @@ use Achsoft\Component\ServiceLocator\Exception\ContainerException;
  *
  * @author Achmad F. Ibrahim <acfatah@gmail.com>
  * @package Achsoft\Component\ServiceLocator\Exception
- * @version 0.1.1
+ * @version 0.1.2
  * @since 0.1.0
  */
 class InvalidIdentifierException extends ContainerException