+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## 0.1.2 (2014.07.06)
+
+* fix #5 Object has to be cloned to remove its reference.
+
 ## 0.1.1 (2014.06.26)
 
-* Add `Container::has` method equivalent to `Container::registered`
-* Rebase git repo
+* Add `Container::has` method equivalent to `Container::registered`.
+* Rebase git repo.
 
 ## 0.1.0 (2014.06.23)
 
-* Initial development
+* Initial development.